digbrain/requirements.txt
# DigBrain Dependencies
# Core
numpy>=1.24.0
scipy>=1.10.0

# Deep Learning
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.0

# Web Framework
aiohttp>=3.9.0
websockets>=12.0

# Database
sqlite3

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.66.0
requests>=2.31.0
aiofiles>=23.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional: For better performance
# flash-attn>=2.0.0  # Requires CUDA
# bitsandbytes>=0.41.0  # For quantization
